+++ src/textual/dom.py
@@ -92,9 +92,6 @@ class DOMNode(MessagePump):
     # Virtual DOM nodes
     COMPONENT_CLASSES: ClassVar[set[str]] = set()
 
-    # Should the content of BINDINGS be treated as priority bindings?
-    PRIORITY_BINDINGS: ClassVar[bool] = False
-
     # Mapping of key bindings
     BINDINGS: ClassVar[list[BindingType]] = []
 
@@ -233,13 +230,13 @@ class DOMNode(MessagePump):
 
         for base in reversed(cls.__mro__):
             if issubclass(base, DOMNode):
-                # See if the current class wants to set the bindings as
-                # priority bindings. If it doesn't have that property on the
-                # class, go with what we saw last.
-                priority = base.__dict__.get("PRIORITY_BINDINGS", priority)
                 if not base._inherit_bindings:
                     bindings.clear()
-                bindings.append(Bindings(base.__dict__.get("BINDINGS", []), priority))
+                bindings.append(
+                    Bindings(
+                        base.__dict__.get("BINDINGS", []),
+                    )
+                )
         keys = {}
         for bindings_ in bindings:
             keys.update(bindings_.keys)